+++ packages/env-utilities/tests/env-file-loader.test.ts
@@ -1,7 +1,7 @@
 import { readdirSync, rmSync, writeFileSync } from 'node:fs';
 import { resolve } from 'node:path';
 
-import { EnvLoaderOptions, loadEnvFiles } from '../src/lib/env-based-loader';
+import { EnvLoaderOptions, loadEnvFiles } from '../src/lib/env-loader';
 
 const originalNodeEnv = process.env.NODE_ENV;
 const fixturesDirectory = resolve(__dirname, 'fixtures');
